fix: 배포 전 정합성 정리 — region/모델 통일 + 하드코딩 제거 + 시연 인프라 최적화
region/모델 ID 정합성:
- k8s 매니페스트(deployment, hpa) AWS_DEFAULT_REGION을 ap-northeast-2 → eu-west-1
- DAG _bedrock_report 모델 ID를 BEDROCK_MODEL_ID 환경변수로 분리, 기본값 sonnet 3.5 v2
- test_bedrock_report 모델 ID 검증값 동기화
- Grafana observability 대시보드 X-Ray 콘솔 링크 region 정정

하드코딩 제거 (CLAUDE.md "하드코딩 금지"):
- DAG/ML/API 상수를 S3_BUCKET/ATHENA_DATABASE/ATHENA_WORKGROUP/ATHENA_OUTPUT 등 env 폴백으로 통일
- alert_handler 메시지 포맷을 if/else 블록으로 정리
- API SAGEMAKER_ENDPOINT_NAME 환경변수화

시연 인프라 최적화 (Terraform):
- Kinesis shard count 변수화(kds_main_shard_count/kds_alert_shard_count) — 시연 시 1로 축소 가능
- Firehose buffer_size/buffer_interval → buffering_size/buffering_interval (AWS provider v5 필드명 변경)
- Firehose dynamic_partitioning false + s3_backup Disabled (시연 단순화)
- Glue bronze 테이블에 Partition Projection 적용 (CLAUDE.md "Partition Projection" 원칙)

기타:
- bronze DDL battery_level/current_load INT → DOUBLE (실제 페이로드와 정합)
- Flink default s3_alerts_path 버킷명을 de-ai-06-smartfactory-bucket로 정정
- generator app.py 시연용 개선 (사용자 IDE 편집)

+++ sql/bronze_ddl.sql
@@ -2,8 +2,8 @@ CREATE EXTERNAL TABLE IF NOT EXISTS robot_telemetry_db.bronze_robot_telemetry (
     robot_id      STRING,
     pos_x         DOUBLE,
     pos_y         DOUBLE,
-    battery_level INT,
-    current_load  INT,
+    battery_level DOUBLE,
+    current_load  DOUBLE,
     motor_temp    DOUBLE,
     `timestamp`   STRING
 )
